# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenbookLM Core API",
    "description": "The public core of OpenbookLM: notebooks, sources, chunks, notes, memories, chat and retrieval metrics. Event-stream semantics are specified separately in docs/contracts/sse-protocol-v1.md.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/memories/{id}": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "GET /api/memories/:id — get a single memory by ID.",
        "operationId": "get_memory_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Memory ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The memory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "memories"
        ],
        "summary": "DELETE /api/memories/:id — delete a single memory.",
        "operationId": "delete_memory_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Memory ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion acknowledgement",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "memories"
        ],
        "summary": "PATCH /api/memories/:id — update a memory's content, salience, or metadata.",
        "operationId": "update_memory_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Memory ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated memory",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/notebooks": {
      "get": {
        "tags": [
          "notebooks"
        ],
        "summary": "GET /api/notebooks - List all notebooks for the current user.",
        "description": "Uses optimized single-query fetch with JOIN to avoid N+1 queries.",
        "operationId": "list_notebooks_handler",
        "responses": {
          "200": {
            "description": "The caller's notebooks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebooksListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "notebooks"
        ],
        "summary": "POST /api/notebooks - Create a new notebook.",
        "operationId": "create_notebook_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotebookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The created notebook",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Denied by the entitlement policy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/notebooks/{id}": {
      "get": {
        "tags": [
          "notebooks"
        ],
        "summary": "GET /api/notebooks/:id - Get a single notebook.",
        "operationId": "get_notebook_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The notebook",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "notebooks"
        ],
        "summary": "DELETE /api/notebooks/:id - Delete a notebook.",
        "description": "Demo notebooks (`is_demo = true`) cannot be deleted.",
        "operationId": "delete_notebook_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion acknowledgement",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "notebooks"
        ],
        "summary": "PATCH /api/notebooks/:id - Update a notebook.",
        "operationId": "update_notebook_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotebookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated notebook",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/notebooks/{id}/chat": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "GET /api/notebooks/{id}/chat - Get paginated chat history.",
        "operationId": "get_chat_history_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of chat history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatHistoryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "POST /api/notebooks/{id}/chat - Send a message and get streaming response.",
        "operationId": "send_message_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Server-sent event stream. Ordering, terminal events and cancellation are specified in docs/contracts/sse-protocol-v1.md, which OpenAPI does not model.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/ChatEvent"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Denied by the entitlement policy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "DELETE /api/notebooks/{id}/chat - Clear chat history.",
        "operationId": "clear_chat_history_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion acknowledgement",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/notebooks/{id}/memories": {
      "get": {
        "tags": [
          "memories"
        ],
        "summary": "GET /api/notebooks/:id/memories — list all memories for a notebook.",
        "description": "Returns memories sorted by salience DESC with plan limit and count.\nValidates that the authenticated user owns the notebook.",
        "operationId": "list_memories_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The notebook's memories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoriesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "memories"
        ],
        "summary": "DELETE /api/notebooks/:id/memories — delete all memories for a notebook.",
        "description": "Validates that the authenticated user owns the notebook.",
        "operationId": "delete_all_memories_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion acknowledgement",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/notebooks/{id}/suggestions": {
      "get": {
        "tags": [
          "suggestions"
        ],
        "operationId": "get_suggestions_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Notebook ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Suggested starter questions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or owned by another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/teaching-modes": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "GET /api/teaching-modes - List available teaching modes.",
        "operationId": "list_teaching_modes",
        "responses": {
          "200": {
            "description": "Available teaching modes and the default",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeachingModesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MemoryResponse": {
        "type": "object",
        "description": "A single memory in the API response.",
        "required": [
          "id",
          "notebook_id",
          "content",
          "memory_type",
          "metadata",
          "salience",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "memory_type": {
            "type": "string"
          },
          "metadata": {},
          "notebook_id": {
            "type": "string",
            "format": "uuid"
          },
          "salience": {
            "type": "number",
            "format": "float"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "UpdateMemoryRequest": {
        "type": "object",
        "description": "Request for updating a memory.",
        "properties": {
          "content": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {},
          "salience": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token verified by the composed identity adapter."
      }
    }
  },
...
}